Split up some long-running regtest modules
