+++ jwst/regtest/test_miri_mrs_spec2.py
@@ -60,51 +60,6 @@ def run_spec2(jail, rtdata_module):
     return rtdata, asn_path
 
 
-@pytest.fixture(scope='module')
-def run_spec3(jail, run_spec2):
-    """Run the Spec3Pipeline on the results from the Spec2Pipeline run"""
-    rtdata, asn_path = run_spec2
-
-    # The presumption is that `run_spec2` has set the input to the
-    # original association. To use this default, and not re-download
-    # the association, simply do not specify `step_params["input_path"]`
-    rtdata.input = asn_path
-    step_params = {
-        'step': 'calwebb_spec3.cfg',
-        'args': [
-            '--steps.master_background.save_results=true',
-            '--steps.mrs_imatch.save_results=true',
-            '--steps.outlier_detection.save_results=true',
-            '--steps.resample_spec.save_results=true',
-            '--steps.cube_build.save_results=true',
-            '--steps.extract_1d.save_results=true',
-            '--steps.combine_1d.save_results=true',
-        ]
-    }
-
-    return rt.run_step_from_dict(rtdata, **step_params)
-
-
-@pytest.fixture(scope='module')
-def run_spec3_multi(jail, rtdata_module):
-    """Run the Spec3Pipeline on multi channel/multi filter data"""
-    step_params = {
-        'input_path': INPUT_PATH + '/' + 'ifushort_set2_asn3.json',
-        'step': 'calwebb_spec3.cfg',
-        'args': [
-            '--steps.master_background.save_results=true',
-            '--steps.mrs_imatch.save_results=true',
-            '--steps.outlier_detection.save_results=true',
-            '--steps.resample_spec.save_results=true',
-            '--steps.cube_build.save_results=true',
-            '--steps.extract_1d.save_results=true',
-            '--steps.combine_1d.save_results=true',
-        ]
-    }
-
-    return rt.run_step_from_dict(rtdata_module, **step_params)
-
-
 @pytest.mark.slow
 @pytest.mark.bigdata
 @pytest.mark.parametrize(
@@ -116,54 +71,6 @@ def test_spec2(run_spec2, fitsdiff_default_kwargs, suffix):
     rtdata, asn_path = run_spec2
     rt.is_like_truth(rtdata, fitsdiff_default_kwargs, suffix,
                      truth_path=TRUTH_PATH)
-
-
-@pytest.mark.slow
-@pytest.mark.bigdata
-@pytest.mark.parametrize(
-    'output',
-    [
-        'ifushort_ch12_spec3_mrs_imatch.fits',
-        'ifushort_ch12_spec3_ch1-medium_s3d.fits',
-        'ifushort_ch12_spec3_ch2-medium_s3d.fits',
-        'ifushort_ch12_spec3_ch1-medium_x1d.fits',
-        'ifushort_ch12_spec3_ch2-medium_x1d.fits',
-    ],
-    ids=["mrs_imatch", "ch1-s3d", "ch2-s3d", "ch1-x1d", "ch2-x1d"]
-)
-def test_spec3(run_spec3, fitsdiff_default_kwargs, output):
-    """Regression test matching output files"""
-    rt.is_like_truth(
-        run_spec3, fitsdiff_default_kwargs, output,
-        truth_path=TRUTH_PATH,
-        is_suffix=False
-    )
-
-
-@pytest.mark.slow
-@pytest.mark.bigdata
-@pytest.mark.parametrize(
-    'output',
-    [
-        'ifushort_set2_0_mrs_imatch.fits',
-        'ifushort_set2_1_mrs_imatch.fits',
-        'ifushort_set2_0_a3001_crf.fits',
-        'ifushort_set2_1_a3001_crf.fits',
-        'ifushort_set2_ch1-short_s3d.fits',
-        'ifushort_set2_ch2-short_s3d.fits',
-        'ifushort_set2_ch1-short_x1d.fits',
-        'ifushort_set2_ch2-short_x1d.fits',
-    ],
-    ids=["ch1-mrs_imatch", "ch2-mrs_imatch", "ch1-crf", "ch2-crf",
-        "ch1-s3d", "ch2-s3d", "ch1-x1d", "ch2-x1d"]
-)
-def test_spec3_multi(run_spec3_multi, fitsdiff_default_kwargs, output):
-    """Regression test matching output files"""
-    rt.is_like_truth(
-        run_spec3_multi, fitsdiff_default_kwargs, output,
-        truth_path=TRUTH_PATH,
-        is_suffix=False
-    )
 
 
 @pytest.mark.slow
